requirements.txt
Flask==2.2.2
tensorflow==2.19.0
mediapipe==0.10.21
spotipy==2.22.0
Pillow==9.0.1
requests==2.28.1
psutil==5.9.1
gunicorn==20.1.0
